@@ -7,13 +7,13 @@ import Foundation
 /// ```
 /// tail -f /tmp/swift_tui_log
 /// ```
-public func log(_ item: Any, terminator: String = "\n") {
+@MainActor public func log(_ item: Any, terminator: String = "\n") {
     print(item, terminator: terminator, to: &logStream)
 }
 
-var logStream = LogStream()
+@MainActor var logStream = LogStream()
 
-struct LogStream: TextOutputStream {
+struct LogStream: TextOutputStream, @unchecked Sendable {
     func write(_ string: String) {
         let logURL = URL(fileURLWithPath: "/tmp/swift_tui_log")
         guard let data = string.data(using: .utf8) else {
