@@ -1,6 +1,6 @@
 import Foundation
 
-class Renderer {
+class Renderer: @unchecked Sendable {
     var layer: Layer
 
     /// Even though we only redraw invalidated parts of the screen, terminal
