@@ -7,11 +7,11 @@ extension Node {
     ///   → VStack<Text>
     ///     → Text
     /// ```
-    func logTree() {
+    @MainActor func logTree() {
         logTree(level: 0)
     }
 
-    private func logTree(level: Int) {
+    @MainActor private func logTree(level: Int) {
         let indent = Array(repeating: " ", count: level * 2).joined()
         log("\(indent)→ \(type(of: self.view))")
         for child in children {
