@@ -41,7 +41,10 @@ class Layer {
             parent.invalidate(rect: Rect(position: rect.position + frame.position, size: rect.size))
             return
         }
-        renderer?.application?.scheduleUpdate()
+        let renderer = self.renderer
+        Task { @MainActor [renderer] in
+            renderer?.application?.scheduleUpdate()
+        }
         guard let invalidated = self.invalidated else {
             self.invalidated = rect
             return
